Initialize global classes explicitly and not at creation time
The use of side effects was quite powerful and dangerous. It happened that this module was imported during the migrator, causing the migrator not taking effect for config changes and possible loss of old config values due to overwriting the default ones in case of special migration.
Now, the classes need to explicitly initialized after class instance creation, to be fully functional. This includes machine as well as the config manager, since the load and validation of config should only happen when the main program starts. Also removed all module level dependencies (e.g. outside function or classes) to the config, since with explicit init, this would not work and use default values instead.