allow virgin only cocktail without all alcoholic ingredients
This will set cocktails as possible where all virgin ingredients meet the available state, without the need of the alcoholic counterparts for the non-virgin recipe.
Alter UI that user directly sees that this is only a virgin one and makes it not possible to select the normal version in this case.